requirements.txt
appdirs==1.4.4
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brotli==1.0.9
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
html5lib==1.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==8.10.0
mpmath==0.0.0
mysql-connector-python==8.4.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.28
ptyprocess==0.7.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
stem==1.8.0
sympy==1.9
systemd-python==234
toripchanger==1.1.3
traitlets==5.1.1
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.18
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
zipp==1.0.0
